backend/pyproject.toml
[tool.poetry]
name = "visitor-management-backend"
version = "0.1.0"
description = "Visitor Management System Backend"
authors = ["Antigravity"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.129.1"
uvicorn = "^0.41.0"
email-validator = "^2.3.0"
python-jose = {extras = ["cryptography"], version = "^3.5.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.22"
bcrypt = "4.3.0"
phonenumbers = "^9.0.24"
sqlalchemy = "^2.0.46"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
